@@ -15,26 +15,22 @@ use Synolia\SyliusSchedulerCommandPlugin\Repository\ScheduledCommandRepositoryIn
 final class LogViewerController extends AbstractController
 {
     /** @var int The time in milliseconds between two AJAX requests to the server. */
-    private $updateTime;
+    private int $updateTime;
 
-    /** @var \Synolia\SyliusSchedulerCommandPlugin\DataRetriever\LogDataRetriever */
-    private $logDataRetriever;
+    private \Synolia\SyliusSchedulerCommandPlugin\DataRetriever\LogDataRetriever $logDataRetriever;
 
-    /** @var \Synolia\SyliusSchedulerCommandPlugin\Repository\ScheduledCommandRepositoryInterface */
-    private $scheduledCommandRepository;
+    private \Synolia\SyliusSchedulerCommandPlugin\Repository\ScheduledCommandRepositoryInterface $scheduledCommandRepository;
 
-    /** @var \Symfony\Contracts\Translation\TranslatorInterface */
-    private $translator;
+    private \Symfony\Contracts\Translation\TranslatorInterface $translator;
 
-    /** @var string */
-    private $logsDir;
+    private string $logsDir;
 
     public function __construct(
         ScheduledCommandRepositoryInterface $scheduledCommandRepository,
         LogDataRetriever $logDataRetriever,
         TranslatorInterface $translator,
         string $logsDir,
-        int $updateTime = 2000
+        int $updateTime = 2000,
     ) {
         $this->logDataRetriever = $logDataRetriever;
         $this->updateTime = $updateTime;
@@ -58,7 +54,7 @@ final class LogViewerController extends AbstractController
                 $this->logsDir . \DIRECTORY_SEPARATOR . $scheduleCommand->getLogFile(),
                 (int) $request->get('lastsize'),
                 (string) $request->get('grep-keywords'),
-                (bool) $request->get('invert')
+                (bool) $request->get('invert'),
             );
 
             return new JsonResponse([
