@@ -12,15 +12,13 @@ use Synolia\SyliusSchedulerCommandPlugin\Entity\CommandInterface;
 
 class SchedulerCommandFixture extends AbstractFixture
 {
-    /** @var RepositoryInterface */
-    private $commandRepository;
+    private \Sylius\Component\Resource\Repository\RepositoryInterface $commandRepository;
 
-    /** @var \Sylius\Component\Resource\Factory\FactoryInterface */
-    private $commandFactory;
+    private \Sylius\Component\Resource\Factory\FactoryInterface $commandFactory;
 
     public function __construct(
         RepositoryInterface $commandRepository,
-        FactoryInterface $commandFactory
+        FactoryInterface $commandFactory,
     ) {
         $this->commandRepository = $commandRepository;
         $this->commandFactory = $commandFactory;
@@ -78,6 +76,7 @@ class SchedulerCommandFixture extends AbstractFixture
             ->booleanNode('executeImmediately')->defaultFalse()->end()
             ->booleanNode('enabled')->defaultTrue()->end()
             ->end()
-            ->end();
+            ->end()
+        ;
     }
 }
