@@ -17,7 +17,7 @@ final class CleanLogAction extends AbstractController
         TranslatorInterface $translator,
         CommandRepository $commandRepository,
         string $command,
-        string $logsDir
+        string $logsDir,
     ): Response {
         /** @var ScheduledCommand|null $scheduleCommand */
         $scheduleCommand = $commandRepository->find($command);
@@ -57,7 +57,7 @@ final class CleanLogAction extends AbstractController
         return $this->redirectToRoute(
             'synolia_admin_command_index',
             [],
-            Response::HTTP_MOVED_PERMANENTLY
+            Response::HTTP_MOVED_PERMANENTLY,
         );
     }
 }
