@@ -9,13 +9,13 @@ use Symfony\Component\Filesystem\Exception\FileNotFoundException;
 final class LogDataRetriever
 {
     /** @var int Maximum amount of bytes this application can load into memory. Default is 2 Megabyte = 2097152 byte */
-    private $maxSizeToLoad = 2097152;
+    private $maxSizeToLoad = 2_097_152;
 
     public function getLog(
         string $filePath,
         int $lastFetchedSize = 0,
         string $grepKeyword = '',
-        bool $invertGrep = false
+        bool $invertGrep = false,
     ): array {
         /** Clear the stat cache to get the latest results */
         \clearstatcache(true, $filePath);
