@@ -27,7 +27,7 @@ class SoftLimitThresholdIsDueCheckerTest extends AbstractIsDueTest
         string $cronExpression,
         \DateTimeInterface $currentDateTime,
         \DateTimeInterface $lastCommandDateTime,
-        bool $expectedResult
+        bool $expectedResult,
     ): void {
         $command = $this->setupCommand($cronExpression);
 
@@ -54,7 +54,7 @@ class SoftLimitThresholdIsDueCheckerTest extends AbstractIsDueTest
     public function testIsDueWithoutScheduledHistory(
         string $cronExpression,
         \DateTimeInterface $currentDateTime,
-        bool $expectedResult
+        bool $expectedResult,
     ): void {
         $command = $this->setupCommand($cronExpression);
 
@@ -71,7 +71,7 @@ class SoftLimitThresholdIsDueCheckerTest extends AbstractIsDueTest
     public function testIsDueUsingCronExpression(
         string $cronExpression,
         \DateTimeInterface $currentDateTime,
-        bool $expectedResult
+        bool $expectedResult,
     ): void {
         $command = $this->setupCommand($cronExpression);
 
