@@ -113,7 +113,8 @@ final class SynoliaSchedulerRunCommandTest extends KernelTestCase
         $command = (new Factory(Command::class))->createNew();
         $command
             ->setCronExpression('* * * * *')
-            ->setCommand('non:existent');
+            ->setCommand('non:existent')
+        ;
 
         $this->save($command);
 
@@ -133,7 +134,7 @@ final class SynoliaSchedulerRunCommandTest extends KernelTestCase
 
         self::assertEquals(
             -1,
-            $persistedCommand->getLastReturnCode()
+            $persistedCommand->getLastReturnCode(),
         );
     }
 
@@ -143,7 +144,8 @@ final class SynoliaSchedulerRunCommandTest extends KernelTestCase
         $command = (new Factory(Command::class))->createNew();
         $command
             ->setCronExpression($cron)
-            ->setCommand('about');
+            ->setCommand('about')
+        ;
 
         return $command;
     }
